refactor: Update import paths for backup and calibrate jobs
